Added assertXML[Not]Equal assertions
This is especially needed to compare XML when hash randomization
is on, as attribute order may vary. Refs #17758.
Thanks Taylor Mitchell for the initial patch.
